refactor: simplify normalizer and remove formatter

+++ testdata/diff/online/add_constraint/diff.sql
@@ -1,2 +1,2 @@
 ALTER TABLE orders
-ADD CONSTRAINT check_amount_positive CHECK (amount > 0);
+ADD CONSTRAINT check_amount_positive CHECK (amount > 0::numeric);
