+++ testdata/diff/online/alter_constraint/diff.sql
@@ -1,4 +1,4 @@
 ALTER TABLE orders DROP CONSTRAINT check_amount_positive;
 
 ALTER TABLE orders
-ADD CONSTRAINT check_amount_positive CHECK (amount > 0);
+ADD CONSTRAINT check_amount_positive CHECK (amount > 0::numeric);
